+++ src/classes/UserPreferences.js
@@ -2,13 +2,14 @@ import { getProp } from "@/helpers/functions/Class";
 
 class UserPreferences {
     constructor() {
-        const localUserPreferences = localStorage.getItem("userPreferences");
+        const localUserPreferences = JSON.parse(localStorage.getItem("userPreferences"));
         this.game = {
             repartition: {
-                forbiddenRoles: getProp(localUserPreferences, "game.repartition.forbidden-roles", []),
+                forbiddenRoles: getProp(localUserPreferences, "game.repartition.forbiddenRoles", []),
                 areRecommendedMinPlayersRespected: getProp(localUserPreferences, "game.repartition.areRecommendedMinPlayersRespected", true),
                 arePowerfulVillagerRolesPrioritized: getProp(localUserPreferences, "game.repartition.arePowerfulVillagerRolesPrioritized", true),
                 arePowerfulWerewolfRolesPrioritized: getProp(localUserPreferences, "game.repartition.arePowerfulWerewolfRolesPrioritized", false),
+                isProTipShown: getProp(localUserPreferences, "game.repartition.isProTipShown", true),
             },
         };
     }