+++ src/plugins/vue-toasted/index.js
@@ -1,4 +1,4 @@
 import Vue from "vue";
 import Toasted from "vue-toasted";
 
-Vue.use(Toasted, { position: "bottom-right", duration: 3000, iconPack: "fontawesome" });
+Vue.use(Toasted, { position: "bottom-right", duration: 3000, iconPack: "fontawesome", keepOnHover: true });